Fix dtype mismatch in quantized LLaDA forward pass

+++ src/chop/nn/quantized/modules/llada/modeling_llada.py
@@ -1841,7 +1841,7 @@ class LLaDAModel(nn.Module):
         if self.config.weight_tying:
             logits = F.linear(x, self.transformer.wte.weight, None)  # type: ignore
         else:
-            logits = self.transformer.ff_out(x)  # type: ignore
+            logits = self.transformer.ff_out(x.to(self.transformer.ff_out.weight.dtype))  # type: ignore
         if self.config.scale_logits:
             logits.mul_(1 / math.sqrt(self.config.d_model))
